Created login, register and main pages and added more queries

+++ untitled/src/main/java/org/example/Manufacturer.java
@@ -1,19 +1,19 @@
 package org.example;
 
 public class Manufacturer {
-    private int id;
+    private Long id;
     private String name;
 
-    public Manufacturer(int id, String name) {
+    public Manufacturer(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
